add tests for Session & Cookie classes
